src/backend/pyproject.toml
[project]
name = "moneta-backend"
version = "0.1.0"
description = "Backend propject steup for Moneta"
readme = "README.md"
requires-python = ">=3.12, <3.13"
dependencies = [
    "semantic-kernel>=1.27.2",
    "opentelemetry-exporter-otlp-proto-grpc>=1.29.0",
    "opentelemetry-instrumentation>=0.50b0",
    "azure-monitor-opentelemetry-exporter>=1.0.0b33",
    "azure-identity>=1.19.0",
    "azure-ai-inference[opentelemetry]>=1.0.0b7",
    "azure-cosmos>=4.9.0",
    "azure-search-documents>=11.5.2",
    "grpcio-tools>=1.68.1",
    "openai>=1.59.2",
    "python-dotenv>=1.0.1",
    "requests>=2.32.3",
    "pyyaml>=6.0.2",
    "fastapi>=0.115.6",
    "uvicorn>=0.32.1",
    "requests-html>=0.10.0",
    "lxml-html-clean>=0.4.1",
    "pandas>=2.2.3",
    "feedparser>=6.0.11"
]
